refactor: move margin and PlatformApi into their own modules
